admin/backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
alembic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
httpx
APScheduler
aiofiles